requirements.txt
google-cloud-bigquery[pandas]
pandas
openpyxl